+++ tests/snapshot_tests/test_snapshots.py
@@ -4,9 +4,11 @@ import pytest
 
 from tests.snapshot_tests.language_snippets import SNIPPETS
 from textual.app import App, ComposeResult
+from textual.binding import Binding
 from textual.containers import Vertical
 from textual.pilot import Pilot
 from textual.widgets import Button, Input, RichLog, TextArea, Footer
+from textual.widgets import Switch
 from textual.widgets.text_area import BUILTIN_LANGUAGES, Selection, TextAreaTheme
 
 # These paths should be relative to THIS directory.
@@ -1527,3 +1529,28 @@ def test_maximize_container(snap_compare):
             yield Footer()
 
     assert snap_compare(MaximizeApp(), press=["m"])
+
+
+def test_check_consume_keys(snap_compare):
+    """Check that when an Input is focused it hides printable keys from the footer."""
+
+    class MyApp(App):
+        BINDINGS = [
+            Binding(key="q", action="quit", description="Quit the app"),
+            Binding(
+                key="question_mark",
+                action="help",
+                description="Show help screen",
+                key_display="?",
+            ),
+            Binding(key="delete", action="delete", description="Delete the thing"),
+            Binding(key="j", action="down", description="Scroll down", show=False),
+        ]
+
+        def compose(self) -> ComposeResult:
+            yield Input(placeholder="First Name")
+            yield Input(placeholder="Last Name")
+            yield Switch()
+            yield Footer()
+
+    assert snap_compare(MyApp())