+++ tests/snapshot_tests/test_snapshots.py
@@ -4,9 +4,11 @@ import pytest
 
 from tests.snapshot_tests.language_snippets import SNIPPETS
 from textual.app import App, ComposeResult
+from textual.binding import Binding
 from textual.containers import Vertical
 from textual.pilot import Pilot
 from textual.widgets import Button, Input, RichLog, TextArea, Footer
+from textual.widgets import Switch
 from textual.widgets.text_area import BUILTIN_LANGUAGES, Selection, TextAreaTheme
 
 # These paths should be relative to THIS directory.
@@ -1536,3 +1538,62 @@ def test_maximize_container(snap_compare):
             yield Footer()
 
     assert snap_compare(MaximizeApp(), press=["m"])
+
+
+def test_check_consume_keys(snap_compare):
+    """Check that when an Input is focused it hides printable keys from the footer."""
+
+    class MyApp(App):
+        BINDINGS = [
+            Binding(key="q", action="quit", description="Quit the app"),
+            Binding(
+                key="question_mark",
+                action="help",
+                description="Show help screen",
+                key_display="?",
+            ),
+            Binding(key="delete", action="delete", description="Delete the thing"),
+            Binding(key="j", action="down", description="Scroll down", show=False),
+        ]
+
+        def compose(self) -> ComposeResult:
+            yield Input(placeholder="First Name")
+            yield Input(placeholder="Last Name")
+            yield Switch()
+            yield Footer()
+
+    assert snap_compare(MyApp())
+
+
+def test_escape_to_minimize(snap_compare):
+    """Regression test for https://github.com/Textualize/textual/issues/4939"""
+
+    TEXT = """\
+    def hello(name):
+        print("hello" + name)
+
+    def goodbye(name):
+        print("goodbye" + name)
+    """
+
+    class TextAreaExample(App):
+        BINDINGS = [("ctrl+m", "screen.maximize")]
+        CSS = """
+        Screen {
+            align: center middle;
+        }
+
+        #code-container {
+            width: 20;
+            height: 10;
+        }
+        """
+
+        def compose(self) -> ComposeResult:
+            with Vertical(id="code-container"):
+                text_area = TextArea.code_editor(TEXT)
+                text_area.cursor_blink = False
+                yield text_area
+
+    # ctrl+m to maximize, escape should minimize
+    assert snap_compare(TextAreaExample(), press=["ctrl+m", "escape"])