+++ tests/test_keys.py
@@ -2,7 +2,7 @@ import pytest
 
 from textual.app import App
 from textual.binding import Binding
-from textual.keys import _character_to_key, format_key
+from textual.keys import _character_to_key, format_key, key_to_character
 
 
 @pytest.mark.parametrize(
@@ -67,3 +67,12 @@ def test_get_key_display():
         == "shift+^]"
     )
     assert app.get_key_display(Binding("delete", "", "")) == "del"
+
+
+def test_key_to_character():
+    assert key_to_character("f") == "f"
+    assert key_to_character("F") == "F"
+    assert key_to_character("space") == " "
+    assert key_to_character("ctrl+space") is None
+    assert key_to_character("question_mark") == "?"
+    assert key_to_character("foo") is None