+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- Added `DOMNode.check_consume_key`
+- Added `DOMNode.check_consume_key` https://github.com/Textualize/textual/pull/4940
 
 ## [0.78.0] - 2024-08-27